Another UPdate

+++ Primes/Generator.cs
@@ -1,15 +1,21 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Numerics;
 using System.Text;
 using System.Threading.Tasks;
 
 namespace Primes
 {
-	//make thread safe? possibly make thread safe alternitive version
+	/// <summary>
+	/// Generator for prime numbers, is not thread safe
+	/// </summary>
     public static class Generator
     {
-		public static IEnumerable<uint> Primes {
+		/// <summary>
+		/// Enumerable with all the primes, brute forces to look for next prime and then compares to all primes less than it's square root.
+		/// </summary>
+		public static IEnumerable<BigInteger> Primes {
 			get
 			{
 				for(int i = 0; ; i++) {
@@ -20,14 +26,18 @@ namespace Primes
 				}
 			}
 		}
-		private static List<uint> generatedPrimes = new List<uint>{ 2, 3, 5, 7, 11 };
+
+		/// <summary>
+		/// Internal storage of generated primes.
+		/// </summary>
+		private static List<BigInteger> generatedPrimes = new List<BigInteger>{ 2, 3, 5, 7, 11 };
 
 		/// <summary>
 		/// Gets n'th prime, considers 2 to be the 0th prime
 		/// </summary>
 		/// <param name="n">which prime to get</param>
 		/// <returns>The nth prime number</returns>
-		public static uint Prime(uint n)
+		public static BigInteger Prime(uint n)
 		{
 			while (generatedPrimes.Count <= n) {
 				generateNextPrime();
@@ -38,7 +48,7 @@ namespace Primes
 		/// <summary>
 		/// Adds 1 number to the generated Primes list
 		/// </summary>
-		private static uint generateNextPrime()
+		private static BigInteger generateNextPrime()
 		{
 			var guess = generatedPrimes[generatedPrimes.Count - 1] + 2;
 			for (; !isPrime(guess); guess += 2) ;
@@ -51,7 +61,7 @@ namespace Primes
 		/// </summary>
 		/// <param name="number">number, must be no greater then the last generated prime squared</param>
 		/// <returns>if numb is prime</returns>
-		private static bool isPrime(uint number)
+		private static bool isPrime(BigInteger number)
 		{
 			for(int i = 0; i * i < number && i < generatedPrimes.Count; i++) {
 				if (number % generatedPrimes[i] == 0)
